ctrl c is wroten

+++ ex1/signals.cpp
@@ -25,6 +25,7 @@ void ctrlZHandler(int sig_num) {
 
   kill(smash.getCurrentPid(), SIGSTOP);
 
+  cout << "smash: process "<<smash.getCurrentPid()<<" was stopped";
 /*pid_t p = fork();
 
 	if (p == 0) {
@@ -48,7 +49,18 @@ void ctrlCHandler(int sig_num) {
 
   SmallShell& smash = SmallShell::getInstance();
 
-  pid_t p = fork();
+  if(smash.getCurrentPid() == 0) {
+    return;
+  }
+
+  if(smash.getJobsList().jobExistsByPID(smash.getCurrentPid())) {
+    smash.getJobsList().removeJobById(smash.getJobsList().jobExistsByPID(smash.getCurrentPid())->getJobID());
+  }
+
+  kill(smash.getCurrentPid(), SIGKILL);
+
+  cout << "smash: process "<<smash.getCurrentPid()<<" was killed";
+/*  pid_t p = fork();
 
 	if (p == 0) {
     smash.executeCommand((string("kill -") + to_string(SIGKILL) + string(" ") + to_string(getppid())).c_str());
@@ -56,11 +68,20 @@ void ctrlCHandler(int sig_num) {
 	}
   else {
     wait(NULL);
-  }
+  }*/
 
 }
 
 void alarmHandler(int sig_num) {
   // TODO: Add your implementation
+
+  cout<<"smash: got an alarm";
+
+  SmallShell& smash = SmallShell::getInstance();
+  
+  /*
+  search which command caused the alarm, send a SIGKILL to its process and print:
+smash: [command-line] timed out!
+  */
 }
